data types

+++ routes/user.py
@@ -20,7 +20,7 @@ def put_user(user:str):
             return JSONResponse(content=content, headers={"Content-Type": "application/json; charset=UTF-8"})
         
 @user_router.post("/submit_score/{uid}/{score}/{leval}")
-def submit_score(uid:str, score:int, leval:int):
+def submit_score(uid:str, score:float, leval:int):
     data = userAuth.update_soce(uid=uid,score=score,leval=leval)
     content = {"data": data}
     return JSONResponse(content=content, headers={"Content-Type": "application/json; charset=UTF-8"})
